% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_children.R
\name{wm_children}
\alias{wm_children}
\alias{wm_children_}
\title{Get children for an AphiaID}
\usage{
wm_children(id, marine_only = TRUE, offset = 1, ...)

wm_children_(id = NULL, name = NULL, marine_only = TRUE, offset = 1,
  ...)
}
\arguments{
\item{id}{(numeric/integer) an AphiaID. For \code{wm_children} it's
required and must be \code{length(id)=1}, for \code{wm_children_} it's
optional and can be \code{length(id) >= 1}}

\item{marine_only}{(logical) marine only or not. default: \code{TRUE}}

\item{offset}{(integer) record to start at. default: 1}

\item{...}{named curl options. see \code{\link[curl]{curl_options}}}

\item{name}{(character) one or more taxonomic names. optional}
}
\description{
Get children for an AphiaID
}
\examples{
wm_children(id = 105706)
wm_children(id = 105706, FALSE)
wm_children(id = 105706, offset = 5)

wm_children(254966)
wm_children(343613)
wm_children_(id = c(254966, 343613))
wm_children_(name = c('Platanista', 'Leucophaeus'))
}
