% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_classification.R
\name{wm_classification}
\alias{wm_classification}
\title{Get classification for an AphiaID}
\usage{
wm_classification(id, marine_only = TRUE, offset = 1, ...)
}
\arguments{
\item{id}{an AphiaID}
}
\description{
Get classification for an AphiaID
}
\examples{
wm_classification(id = 105706)
}

