% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_common_id.R
\name{wm_common_id}
\alias{wm_common_id}
\title{Get vernacular names from an AphiaID}
\usage{
wm_common_id(id, ...)
}
\arguments{
\item{id}{an AphiaID}
}
\description{
Get vernacular names from an AphiaID
}
\examples{
wm_common_id(id = 105706)
wm_common_id(id = 156806)
wm_common_id(id = 397065)
}

