% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_distribution.R
\name{wm_distribution}
\alias{wm_distribution}
\title{Get distribution data by AphiaID}
\usage{
wm_distribution(id, ...)
}
\arguments{
\item{x}{xxx}
}
\description{
Get distribution data by AphiaID
}
\examples{
wm_distribution(id = 1080)
}

