% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_externalid.R
\name{wm_external}
\alias{wm_external}
\title{Get AphiaID via an external ID}
\usage{
wm_external(id, type = "tsn", ...)
}
\arguments{
\item{type}{(character) the type of external id. one of: tsn, bold,
dyntaxa, eol, fishbase, iucn, lsid, ncbi, gisd. default: tsn}

\item{x}{id an external ID}
}
\description{
Get AphiaID via an external ID
}
\examples{
wm_external(id = 1080)
wm_external(id = 105706)

library(crul)
wm_external(id = 105706, verbose = TRUE)
}

