% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_name2id.R
\name{wm_name2id}
\alias{wm_name2id}
\title{Get AphiaID from a taxonomic name}
\usage{
wm_name2id(name, ...)
}
\arguments{
\item{name}{(character) a taxon name}

\item{...}{named curl options. see \code{\link[curl]{curl_options}}}
}
\description{
Get AphiaID from a taxonomic name
}
\examples{
wm_name2id(name = "Rhincodon")
}
