% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_record.R
\name{wm_record}
\alias{wm_record}
\title{Get complete AphiaRecord for an AphiaID}
\usage{
wm_record(id, ...)
}
\arguments{
\item{id}{(numeric/integer) an AphiaID. required.}

\item{...}{named curl options. see \code{\link[curl]{curl_options}}}
}
\description{
Get complete AphiaRecord for an AphiaID
}
\examples{
wm_record(id = 105706)
}
