% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_record.R
\name{wm_record_}
\alias{wm_record_}
\title{Get complete AphiaRecord for an AphiaID - plural version}
\usage{
wm_record_(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Get complete AphiaRecord for an AphiaID - plural version
}
\details{
This function is defunct - see \code{\link[=wm_record]{wm_record()}}
}
