% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_records_taxamatch.R
\name{wm_records_taxmatch}
\alias{wm_records_taxmatch}
\title{Get records for onen or more taxonomic name(s) using
the TAXAMATCH fuzzy matching algorithm}
\usage{
wm_records_taxmatch(name, marine_only = TRUE, ...)
}
\arguments{
\item{name}{(character) start date. required.}

\item{marine_only}{(logical) marine only or not. default: \code{TRUE}}

\item{...}{named curl options. see \code{\link[curl]{curl_options}}}
}
\description{
Get records for onen or more taxonomic name(s) using
the TAXAMATCH fuzzy matching algorithm
}
\examples{
wm_records_taxmatch(name = 'Platanista gangetica')
wm_records_taxmatch(name = c('Platanista gangetica', 'Coryphaena'))
}

