% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_sources.R
\name{wm_sources}
\alias{wm_sources}
\title{Get sources for an AphiaID}
\usage{
wm_sources(id, ...)
}
\arguments{
\item{id}{(numeric/integer) an AphiaID. required.}

\item{...}{named curl options. see \code{\link[curl]{curl_options}}}
}
\description{
Get sources for an AphiaID
}
\examples{
wm_sources(id = 105706)
}

