% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_synonyms.R
\name{wm_synonyms}
\alias{wm_synonyms}
\title{Get synonyms for an AphiaID}
\usage{
wm_synonyms(id, ...)
}
\arguments{
\item{id}{(numeric/integer) an AphiaID. required.}

\item{...}{named curl options. see \code{\link[curl]{curl_options}}}
}
\description{
Get synonyms for an AphiaID
}
\examples{
wm_synonyms(id = 105706)
}
