% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worrms-package.R
\docType{package}
\name{worrms-package}
\alias{worrms}
\alias{worrms-package}
\title{World Register of Marine Species Client}
\description{
World Register of Marine Species Client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

