% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_children.R
\name{wm_children}
\alias{wm_children}
\title{Get children for an AphiaID}
\usage{
wm_children(id, marine_only = TRUE, offset = 1, ...)
}
\arguments{
\item{id}{an AphiaID}

\item{marine_only}{(logical) marine only or not. default: \code{TRUE}}

\item{offset}{(integer) record to start at. default: 1}
}
\description{
Get children for an AphiaID
}
\examples{
wm_children(id = 105706)
wm_children(id = 105706, FALSE)
wm_children(id = 105706, offset = 5)
}

