% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_classification.R
\name{wm_classification}
\alias{wm_classification}
\alias{wm_classification_}
\title{Get classification for an AphiaID}
\usage{
wm_classification(id, ...)

wm_classification_(id = NULL, name = NULL, ...)
}
\arguments{
\item{id}{(numeric/integer) an AphiaID. For \code{wm_children} it's
required and must be \code{length(id)=1}, for \code{wm_children_} it's
optional and can be \code{length(id) >= 1}}

\item{...}{named curl options. see \code{\link[curl]{curl_options}}}

\item{name}{(character) one or more taxonomic names. optional}
}
\description{
Get classification for an AphiaID
}
\examples{
wm_classification(id = 105706)
wm_classification(id = 126436)

wm_classification(254967)
wm_classification(344089)
wm_classification_(id = c(254967, 344089))
wm_classification_(name = c('Platanista gangetica', 'Leucophaeus scoresbii'))
}
