% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_classification.R
\name{wm_classification}
\alias{wm_classification}
\title{Get classification for an AphiaID}
\usage{
wm_classification(id, ...)
}
\arguments{
\item{id}{(numeric/integer) an AphiaID. required.}

\item{...}{named curl options. see \code{\link[curl]{curl_options}}}
}
\description{
Get classification for an AphiaID
}
\examples{
wm_classification(id = 105706)
}

