% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_distribution.R
\name{wm_distribution}
\alias{wm_distribution}
\title{Get distribution data by AphiaID}
\usage{
wm_distribution(id, ...)
}
\arguments{
\item{id}{an AphiaID. required.}

\item{...}{named curl options. see \code{\link[curl]{curl_options}}}
}
\description{
Get distribution data by AphiaID
}
\examples{
wm_distribution(id = 156806)
}

