% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_id2name.R
\name{wm_id2name}
\alias{wm_id2name}
\title{Get taxonomic name for an AphiaID}
\usage{
wm_id2name(id, ...)
}
\arguments{
\item{id}{(numeric/integer) an AphiaID. required.}

\item{...}{named curl options. see \code{\link[curl]{curl_options}}}
}
\description{
Get taxonomic name for an AphiaID
}
\examples{
wm_id2name(id = 105706)
}

