% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_record.R
\name{wm_record}
\alias{wm_record}
\alias{wm_record_}
\title{Get complete AphiaRecord for an AphiaID}
\usage{
wm_record(id, ...)

wm_record_(id = NULL, name = NULL, ...)
}
\arguments{
\item{id}{(numeric/integer) an AphiaID. For \code{wm_record} it's
required and must be \code{length(id) == 1}, for \code{wm_record_} it's
optional and can be \code{length(id) >= 1}}

\item{...}{named curl options. see \code{curl::curl_options}}

\item{name}{(character) one or more taxonomic names. optional}
}
\value{
A named list. When using underscore method, each output is named
by the input ID, and can be separated by the list names
}
\description{
Get complete AphiaRecord for an AphiaID
}
\note{
\code{wm_record_} is defunct, \code{wm_record} can do plural requests now
}
\section{Singular vs. plural}{

Of the two sister functions, the one without the underscore is the original
function that wraps the relavant WoRMS API method - and only accepts
one thing (i.e., name or AphiaID) per request.

The sister function with the underscore at the end is the plural version,
accepting more than one input. Internally this function loops over
the non-underscore method, and labels output (whether it's a list or
data.frame rows) with the input names or IDs so that you can easily
parse output by your inputs.
}

\examples{
\dontrun{
wm_record(id = 105706)
wm_record(id = c(105706, 126436))
wm_record_(id = c(105706, 126436))
}
}
