% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_record.R
\name{wm_record}
\alias{wm_record}
\title{Get complete AphiaRecord for an AphiaID}
\usage{
wm_record(id, ...)
}
\arguments{
\item{id}{(numeric/integer) one or more AphiaID's.}

\item{...}{named curl options. see \code{curl::curl_options}}
}
\value{
A data.frame, one record for each id
}
\description{
Get complete AphiaRecord for an AphiaID
}
\note{
\code{wm_record_} is defunct, \code{wm_record} can do plural requests now
}
\examples{
\dontrun{
wm_record(id = 105706)
wm_record(id = c(105706, 126436))
}
}
