% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_record_by_external.R
\name{wm_record_by_external}
\alias{wm_record_by_external}
\title{Get record by external ID}
\usage{
wm_record_by_external(id, type = "tsn", ...)
}
\arguments{
\item{id}{(numeric/integer) an AphiaID. required.}

\item{type}{(character) the type of external id. one of: tsn, bold,
dyntaxa, eol, fishbase, iucn, lsid, ncbi, gisd. default: tsn}

\item{...}{named curl options. see \code{\link[curl]{curl_options}}}
}
\description{
Get record by external ID
}
\examples{
wm_record_by_external(id = 85257)
wm_record_by_external(id = 159854)
}
