% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worrms-package.R
\docType{package}
\name{worrms-package}
\alias{worrms}
\alias{worrms-package}
\title{World Register of Marine Species Client}
\description{
World Register of Marine Species Client
}
\section{Fail behavior}{

The WoRMS REST API doesn't have sophisticated error messaging, so
most errors will result in a \code{(204) - No Content} or
in \code{(400) - Bad Request}

Because WoRMS doesn't do comprehensive error reporting, we do a fair
amount of checking user inputs to help prevent errors that will be
meaningless to the user. Let us know if we can improve on this.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

